Build script improvements
