+++ .github/workflows/gradle.yml
@@ -7,11 +7,14 @@ on: [push, pull_request]
 
 jobs:
   build:
+    # Only run on PRs if the source branch is on a different repo. We do not need run everything twice.
+    if: ${{ github.event_name != 'pull_request' || github.repository != github.event.pull_request.head.repo.full_name }}
 
     runs-on: ubuntu-latest
 
     steps:
     - uses: actions/checkout@v2
+    - uses: gradle/wrapper-validation-action@v1
     - name: Set up JDK 11
       uses: actions/setup-java@v2
       with:
