+++ adventure/build.gradle.kts
@@ -1,5 +1,3 @@
-import com.github.jengelman.gradle.plugins.shadow.tasks.ShadowJar
-
 plugins {
     id("com.github.johnrengelman.shadow")
 }
@@ -7,11 +5,11 @@ plugins {
 // Shade and relocate adventure in an extra module, so that common/the rest can directly depend on a
 // relocated adventure without breaking native platform's adventure usage with project wide relocation
 tasks {
-    withType<ShadowJar> {
+    shadowJar {
         relocate("net.kyori", "com.viaversion.viaversion.libs.kyori")
     }
-    getByName("build") {
-        dependsOn(withType<ShadowJar>())
+    build {
+        dependsOn(shadowJar)
     }
 }
 
@@ -23,3 +21,5 @@ dependencies {
         exclude("com.google.code.gson", "gson")
     }
 }
+
+publishShadowJar()
