+++ build-logic/src/main/kotlin/via.base-conventions.gradle.kts
@@ -1,0 +1,42 @@
+plugins {
+    `java-library`
+    `maven-publish`
+}
+
+tasks {
+    // Variable replacements
+    processResources {
+        filesMatching(listOf("plugin.yml", "mcmod.info", "fabric.mod.json", "bungee.yml")) {
+            expand("version" to project.version, "description" to project.description)
+        }
+    }
+    javadoc {
+        options.encoding = Charsets.UTF_8.name()
+    }
+    compileJava {
+        options.encoding = Charsets.UTF_8.name()
+        options.compilerArgs.addAll(listOf("-nowarn", "-Xlint:-unchecked", "-Xlint:-deprecation"))
+    }
+}
+
+java {
+    javaTarget(8)
+    withSourcesJar()
+    withJavadocJar() // todo: do we need javadoc enabled for every module?
+}
+
+publishing {
+    publications.create<MavenPublication>("mavenJava") {
+        groupId = rootProject.group as String
+        artifactId = project.name
+        version = rootProject.version as String
+    }
+    repositories.maven {
+        name = "Via"
+        url = uri("https://repo.viaversion.com/")
+        credentials(PasswordCredentials::class)
+        authentication {
+            create<BasicAuthentication>("basic")
+        }
+    }
+}
