+++ build-logic/src/main/kotlin/via.standard-conventions.gradle.kts
@@ -1,0 +1,5 @@
+plugins {
+    id("via.base-conventions")
+}
+
+publishJavaComponents()
