+++ gradle.properties
@@ -1,3 +1,7 @@
+# Project properties - we put these here so they can be modified without causing a recompile of the build scripts
+projectVersion=4.0.0-21w20a-SNAPSHOT
+
+# Gradle properties
 org.gradle.daemon=true
 org.gradle.configureondemand=true
 org.gradle.parallel=true
